refactor: extract preview server helpers
